# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "مسیر کانکور — API",
    "description": "API بک‌اند سایت آموزشگاه کانکور. اکثر مسیرها نیازمند توکن JWT هستند (هدر Authorization: Bearer <token>).",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "نسبی — همان دامنه‌ی فرانت‌اند (پشت پروکسی/nginx)" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "description": "نقش کاربر اجازه‌ی این عملیات را نمی‌دهد.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "منبع مورد نظر یافت نشد.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/teachers": {
      "get": {
        "tags": ["teachers"],
        "summary": "لیست استادان (فقط ادمین)",
        "responses": { "200": { "description": "OK" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["teachers"],
        "summary": "ساخت حساب استاد جدید (فقط ادمین)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["full_name", "email", "password"],
                "properties": { "full_name": { "type": "string" }, "email": { "type": "string" }, "password": { "type": "string", "minLength": 6 } }
              }
            }
          }
        },
        "responses": { "201": { "description": "ساخته شد" }, "409": { "description": "ایمیل تکراری" } }
      }
    },
...
    "/forum/course/{courseId}": {
      "get": {
        "tags": ["forum"],
        "summary": "موضوعات انجمن یک کورس",
        "parameters": [{ "name": "courseId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "post": {
        "tags": ["forum"],
        "summary": "ساخت موضوع جدید در انجمن کورس",
        "parameters": [{ "name": "courseId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["title"], "properties": { "title": { "type": "string", "maxLength": 200 }, "body": { "type": "string" } } } } }
        },
        "responses": { "201": { "description": "ساخته شد" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/forum/thread/{id}": {
      "get": {
        "tags": ["forum"],
        "summary": "جزئیات یک موضوع انجمن + پاسخ‌ها",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
    "/forum/thread/{id}/reply": {
      "post": {
        "tags": ["forum"],
        "summary": "پاسخ به یک موضوع انجمن",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["message"], "properties": { "message": { "type": "string", "maxLength": 5000 } } } } }
        },
        "responses": { "201": { "description": "ثبت شد" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
  }
}